Update WebVR 1.1 spec to account for multi-GPU handling
Following the language from the PR for the 2.0 explainer (https://github.com/w3c/webvr/pull/226), this adds a paragraph that describes the expected sequence of events during requestPresent, in the event that a WebGL context is not compatible with the VRDisplay's graphics adapter. The webglcontextlost will be fired, and if Event.preventDefault() is called, the user agent will at that point switch to the VRDisplay's graphics adapter and continue the requestPresent process.