Add polygonal stage bounds.
Addresses #228. I’m proposing keeping the bounds rect in the spec,
though slightly modified to make use of the new point type, because I
believe that it’s going to be the more generally useful value for
developers. It’s true that the same value could be computed from the
bounds geometry, but the methods for doing that is not completely
trivial and possibly something better handled by the UA for consistency
(also, the native APIs may provide it anyway).

Even if we drop the play area rect, though I’d like to keep the
distinction between the bounds object and the geometry array to allow
for future extension.