Rationalizing and clarifying XRFrameOfReference and XRCoordinateSystem (#409)
Addresses most of the confusion and concerns discussed at the Sept `18 F2F regarding tracking systems, the purpose of frames of references, and their relationship to coordinate systems. It also addresses issue #396, issue #389, issue #367, issue #355, and supersedes PR #358. This change does not address issue #384 and issue #403, though it will impact the approach to solving them as well.  It's also worth pointing out that because we haven't officially agreed on whether or not XRAnchor should be part of the core of WebXR 1.0, there is relatively little reference the concept in this new document as of yet.

The frame of reference types are now: `XRBoundedFrameOfReference`, `XRUnboundedFrameOfReference`, and `XRStationaryFrameOfReference`.  The latter has three subtypes: `floor-level`, `eye-level`, and `position-disabled`.  The unified rendering path is now supported even when an XR device is not present by allowing the `getDevicePose()` and `getInputPose()` to accept a null frame of reference.