Add ignoreDepthValues attribute to the XRWebGLLayer (#548)
This value instructs the UA to prevent the values left in the depth
buffer after a frame is rendered for compositing. If left false (the
default) the UA is allowed to use the depth buffer for things like
improved reprojection.

Also includes some language specifying how the depth buffer is
interpreted and indicating that reverse-Z projection is acceptable.