Remove viewMatrix and add XRRigidTransform.inverse() (#531)
Long-requested API cleanup to get rid of the explicit `viewMatrix` attribute, paired with a method for getting the inverse of an `XRRigidTransform`. The `inverse()` method can both be used to get the value that was previously returned by the `viewMatrix` (`transform.inverse().matrix`) and also makes application of the `XRReferenceSpace.originOffset` attribute easier if your use case happens to feel "backwards" from the default behavior.