Indicate when compositing is using depth values
This addresses a concern brought up by @kearwood during the discussion
of #523. He pointed out that it would be useful for some apps to know
if the compositor actually was using the depth values in the case where
`ignoreDepthValues` was not set to `true` at layer creation time,
because that would allow some applications to skip an unnecessary
explicit depth write step that they otherwise would be happy to do if
they knew it would contribute to the final frame.

Following the pattern set by other WebGL-style attributes like
`antialias`, this PR indicates that the layer attribute should
communicate whether or not depth values are actually being ignored by
the compositor, rather than simply echo the boolean passed in at layer
creation time.