Spec: Move outputContext to XRRenderState (#536)
This change makes the outputContext part of the Session's render state, as is outlined in the explainer, so that it can be managed by the same timing as the rest of the state. This change requires slightly more logic than other render state values since it may need to remove the outputContext from a session that it was previously assigned to.