Specify that projection matrices may include shear
Fixes #461.

At the January F2F we agreed that the `projectionMatrix` attribute
should remain a matrix to allow for more complex projections that
include, for example, shearing. In order to dissuade applications from
attempting to decompose the matrix into a view frustum, which wouldn't
properly account for that, this PR adds text to the spec that explicitly
states that decomposition is a bad idea and gives reasoning as to why.