Spec: Gamepad Integration (#553)
Specifies how the WebXR Device API makes use of the Gamepad API. Specifically, defines a `gamepad` attribute on the `XRInputSource` that represents the button and axis state of a controller device. Also describes a new standard mapping (`xr-standard`) to describe layouts of common XR controller buttons and axes as relates to the `gamepad` value arrays.