Consolidate reference space types and interfaces (#587)
Consolidate the various interfaces that extended XRReferenceSpace, with the exception of XRBoundedReferenceSpace which was still necessary to communicate the boundsGeometry. This PR also eliminates the 'stationary' XRReferenceSpaceType in favor of making all of it's subtypes top level types. Which, in turn, eliminates the need for the options dictionary when creating a reference space, so this PR also changes that call to only take the type enum (an options dictionary can be added as a second argument in the future if needed, as described by #542.