Simplify available spaces (#626)
Removes the position-disabled reference space from the spec and
explainers. Developers that wish to display the types of content this
reference space supported (primarily 360 media) should instead manually
zero-out the position of the XRView transforms when rendering.

Also consolidates the viewerSpace attribute and the identity reference
space into a single new reference space called viewer. This new
space is functionally equivalent to the previous viewerSpace but
fetched via requestReferenceSpace() with the same guarantee of
availability as identity had previously. All other spaces relate to it
like any other space.