Change inputSources getter from method to attrib. (#624)
Changes the `getInputSources()` method on `XRSesssion` to a readonly
attribute `inputSources`. This better reflects the expected behavior
of the returned array being a live object that is updated in-place as
input sources are added and removed.

In order to facilitate that change the `XRInputSourceArray` interface was
also added, since `FrozenArray` doesn't support the live changes that are
required for this attribute.