Single-canvas inline, drop XRPresentationContext (#656)
This change adds a path for using a single WebGL canvas to drive inline
sessions, rather than needing both a WebGL context and an
XRPresentationContext before the inline session can function. With this
change `XRWebGLLayer`s will automatically stop allocating a framebuffer
with inline sessions, which causes WebGL commands to naturally "fall through"
to the underlying default framebuffer.

In addition to the above, this change removes the XRPresentationContext
entirely, and along with it the ability to do desktop mirroring (and
more advanced inline scenarios.) This is to simplify the core API, and
equivalent functionality will be reconsidered in the future because the
functionality that they represent is valuable.