Initial explainer text for privacy/security aspects of WebXR (#733)
An explainer document has been added that will eventually define and address the privacy and security challenges we are attempting to solve. For right now, this document is more outline than filled in content, and is filled with a lot of TODOs that reference the GitHub issues covering each individual topic. This change gets the bones of the document in place so it will be easier to examine each subsequent change as a diff on top of a solid base.