Made the xr-standard gamepad mapping more rigid (#735)
This will make cross-platform implementations more compatible by
removing some level of platform option from the mapping process. It will
also make it more practical to expose the discussed "profiles" concept
by making it easier to have inter-compatible sets of profiles.

We've also realized that the presence of profiles makes the requirements
around what elements must be present on an `xr-standard` device less
necessary, and as such have loosened that requirement to indicate that
only a primary button (separate from a touchpad or thumbstick) be present.

Finally, took the opportunity to normalized on the verbiage "thumbstick"
instead of joystick, since that appears to be the official verbiage used
by all the device vendors.