Require that sensitive UI hides WebXR content (#742)
/fixes 724

Until we get some better guidelines in place regarding what kind of
mitigations are appropriate for using the `visible-blurred` state when
sensitive UI is being displayed, simply require that the content is
hidden instead. `visible-blurred` may still be used in less sensitive
cases where the XR content is not in the foreground but still visible.