Enable optional/required features to be checked before session creation (#739)
This change describes the mechanism by which developers can block session creation if a feature they absolute require is unavailable. While this may occasionally have to do with consent, it also may simply be due to lack of hardware support. Following the same pattern, this design allows developers to indicate up front that there are optional features they would very much like to use. This allows any necessary explicit consent (for both optional and required features) to be requested before a session begins. For now, the list of features only includes reference space features, but we've heard feedback from folks working on new functionality that a similar system is needed.